requirements.txt
python-telegram-bot==20.3
python-telegram-bot[socks]==20.3
requests